+++ packages/theme/docs/test/abc/dynamic/[version].paths.js
@@ -1,0 +1,6 @@
+export default {
+  paths: () => [
+    { params: { title: '动态路由1', description: 'foo1 description', version: '1.0.0' }, content: '动态渲染内容1' },
+    { params: { title: '动态路由2', description: 'foo1 descriptionxx', version: '2.0.0' }, content: '## 测试内容' },
+  ]
+}