+++ packages/theme/src/node.ts
@@ -46,7 +46,7 @@ export function getThemeConfig(cfg: Partial<Theme.BlogConfig> = {}) {
   return {
     themeConfig: {
       blog: {
-        pagesData,
+        pagesData, // 插件里补全
         ...cfg
       },
       // 补充一些额外的配置用于继承
@@ -61,6 +61,10 @@ export function getThemeConfig(cfg: Partial<Theme.BlogConfig> = {}) {
  */
 export function defineConfig(config: UserConfig<Theme.Config>): any {
   return config
+}
+
+export function defineLocaleConfig(cfg: Omit<Theme.BlogConfig, 'locales' | 'pagesData'>) {
+  return cfg
 }
 
 // 重新导包 tabsMarkdownPlugin 导出CJS格式支持