+++ packages/theme/src/utils/node/theme.ts
@@ -1,9 +1,6 @@
-/* eslint-disable prefer-rest-params */
 import fs from 'node:fs'
 import path from 'node:path'
-import process from 'node:process'
-import glob from 'fast-glob'
-import { getDefaultTitle, getFileLastModifyTime, getTextSummary, grayMatter, normalizePath } from '@sugarat/theme-shared'
+import { getDefaultTitle, getFileLastModifyTime, getTextSummary, getVitePressPages, grayMatter, normalizePath, renderDynamicMarkdown } from '@sugarat/theme-shared'
 import type { SiteConfig } from 'vitepress'
 import type { Theme } from '../../composables/config/index'
 import { formatDate } from '../client'
@@ -29,8 +26,8 @@ export function getPageRoute(filepath: string, srcDir: string) {
 }
 
 const defaultTimeZoneOffset = new Date().getTimezoneOffset() / -60
-export async function getArticleMeta(filepath: string, route: string, timeZone = defaultTimeZoneOffset) {
-  const fileContent = await fs.promises.readFile(filepath, 'utf-8')
+export async function getArticleMeta(filepath: string, route: string, timeZone = defaultTimeZoneOffset, baseContent?: string) {
+  const fileContent = baseContent || await fs.promises.readFile(filepath, 'utf-8')
 
   const { data: frontmatter, excerpt, content } = grayMatter(fileContent, {
     excerpt: true,
@@ -70,6 +67,7 @@ export async function getArticleMeta(filepath: string, route: string, timeZone =
     = meta.description || getTextSummary(content, 100) || excerpt
 
   // 获取封面图
+  // TODO: 耦合信息优化
   meta.cover
     = meta.cover
     ?? (getFirstImagURLFromMD(fileContent, route))
@@ -81,19 +79,18 @@ export async function getArticleMeta(filepath: string, route: string, timeZone =
   }
   return meta as Theme.PageMeta
 }
-export async function getArticles(cfg: Partial<Theme.BlogConfig>, vpConfig: SiteConfig) {
-  const srcDir
-    = cfg?.srcDir || vpConfig.srcDir.replace(vpConfig.root, '').replace(/^\//, '')
-    || process.argv.slice(2)?.[1]
-    || '.'
-  const files = glob.sync(`${srcDir}/**/*.md`, { ignore: ['node_modules'], absolute: true })
 
-  const metaResults = files.reduce((prev, curr) => {
-    const route = getPageRoute(curr, vpConfig.srcDir)
-    const metaPromise = getArticleMeta(curr, route, cfg?.timeZone)
+export async function getArticles(cfg: Partial<Theme.BlogConfig>, vpConfig: SiteConfig) {
+  const pages = getVitePressPages(vpConfig)
+  const metaResults = pages.reduce((prev, value) => {
+    const { page, route, originRoute, filepath, isDynamic, dynamicRoute } = value
+
+    const metaPromise = (isDynamic && dynamicRoute)
+      ? getArticleMeta(filepath, originRoute, cfg?.timeZone, renderDynamicMarkdown(filepath, dynamicRoute.params, dynamicRoute.content))
+      : getArticleMeta(filepath, originRoute, cfg?.timeZone)
 
     // 提前获取，有缓存取缓存
-    prev[curr] = {
+    prev[page] = {
       route,
       metaPromise
     }
@@ -105,8 +102,8 @@ export async function getArticles(cfg: Partial<Theme.BlogConfig>, vpConfig: Site
 
   const pageData: Theme.PageData[] = []
 
-  for (const file of files) {
-    const { route, metaPromise } = metaResults[file]
+  for (const page of pages) {
+    const { route, metaPromise } = metaResults[page.page]
     const meta = await metaPromise
     if (meta.layout === 'home') {
       continue
@@ -142,5 +139,5 @@ export function patchVPThemeConfig(
 }
 
 export function checkConfig(cfg?: Partial<Theme.BlogConfig>) {
-  // TODO：保留
+  // 保留
 }