+++ packages/theme/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @sugarat/theme
 
+## 0.5.0
+
+### Minor Changes
+
+- feat: inline group icons plugin
+- feat: support i18n
+- feat: support rewrites and dynamic routes
+- chore: Optimization of theme reloading logic in dev mode
+- Updated dependencies
+  - @sugarat/theme-shared@0.0.3
+  - vitepress-plugin-pagefind@0.4.11
+  - vitepress-plugin-rss@0.3.0
+  - vitepress-plugin-announcement@0.1.3
+
 ## 0.4.13
 
 ### Patch Changes